xview2_strong_baseline/requirements.txt
# Edits:
# - Commented out Hydra==2.5 as it seems to be incorrectly referring to the wrong package


torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.0.2
torchvision==0.15.2
lightning==2.0.6
scikit-learn==1.3.0
shapely==2.0.1
opencv-python==4.8.0.74
omegaconf==2.3.0
#Hydra==2.5
hydra-colorlog==1.2.0
hydra-core==1.3.2
wandb==0.15.8